agents/pyproject.toml
[tool.poetry]
name = "sentinel-agents"
version = "0.1.0"
description = "SENTINEL Multi-Agent Audit System"
authors = ["motafegh"]
package-mode = false
[tool.poetry.dependencies]
python = ">=3.11,<3.15"

# LangChain ecosystem — current 2026 stack
langchain = "^0.3"
langchain-openai = "^0.2"        # OpenAI-compatible — works with LM Studio
langchain-community = "^0.3"     # FAISS, BM25, document loaders
langgraph = "^0.2"               # Agent graph orchestration

# RAG
faiss-cpu = "^1.8"               # Vector similarity search
rank-bm25 = "^0.2"               # BM25 keyword search for hybrid RAG

# Structured outputs
pydantic = "^2.0"                # Type-safe schemas

# MCP
mcp = "^1.0"                     # Model Context Protocol

# HTTP + utils
httpx = "^0.27"                  # async HTTP for Module 1 API calls
loguru = "^0.7"                  # consistent logging with rest of project
python-dotenv = "^1.0"           # load .env
filelock = "^3.13"               # FIX-8: single-writer lock for index files — prevents concurrent pipeline corruption
dagster = "^1.12.22"
dagster-webserver = "^1.12.22"
web3 = "^7.15.0"
torch = "^2.11.0"
numpy = "^2.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"

[tool.pytest.ini_options]
# Override root sentinel/pyproject.toml which has --cov addopts (pytest-cov
# not installed in agents venv — tests would fail with "unrecognized arguments").
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts      = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"